added tspec infer types
